wip: first draft of new tasks/inputTasks

+++ implementation/jvm/src/main/scala/com/github/tammo/yabt/task/jvm/tasks/CleanTask.scala
@@ -3,18 +3,18 @@ package com.github.tammo.yabt.task.jvm.tasks
 import com.github.tammo.yabt.task.Task
 import com.github.tammo.yabt.task.TaskDSL.task
 
-import java.nio.file.{Files, Paths}
+import java.nio.file.{Files, Path}
 import java.util.Comparator
 
 object CleanTask:
 
   lazy val cleanTask: Task[Unit] =
     task("clean", "Cleans all generated files.") { context =>
-      val targetDirectory =
-        Paths.get(context.workingDirectory.toString, "target")
+      val targetDirectory = context.moduleDirectory.resolve("target")
 
-      Files
-        .walk(targetDirectory)
-        .sorted(Comparator.reverseOrder())
-        .forEach(Files.deleteIfExists(_))
+      if (Files.exists(targetDirectory))
+        Files
+          .walk(targetDirectory)
+          .sorted(Comparator.reverseOrder())
+          .forEach(Files.deleteIfExists(_))
     }
